\encoding{utf8}
\docType{data}
\name{dentist}
\alias{dentist}
\title{Dentist diagnosis data set}
\format{A data matrix with 3869 rows and 5 columns.}
\description{
The Handelman’s dentistry data (Handelman et al. , 1986) display the evaluation of 3869 dental
x-rays that may show incipient caries performed by five dentists.
}
\examples{
data("dentist")
}
\references{
Handelman, S.L., Leverett, D.H., Espeland, M.A., & Curzon, J.A. 1986. Clinical radiographic
evaluation of sealed carious and sound tooth surfaces. The Journal of the American Dental
Association, 113(5), 751–754.
}
\keyword{datasets}

