\name{parameterscat-class}
\alias{parameterscat-class}
\title{parameterscat output objects}
\description{
  This class contains the parameters estimated by the
  clustercat algorithm for a fixed number of class.
  \describe{ \item{rho}{List where rho[[k]][[b]] indicated
  the proportion of the maximum dependence distribution in
  the block b of the class k.} \item{tau}{List where
  tau[[k]][[b]] indicated parameters of the first block
  variable multinomial of the maximum dependence
  distribution in the block b of the class k.}
  \item{delta}{List where delta[[k]][[b]] indicated
  modalities relations involving by the maximum dependence
  distribution in the block b of the class k.}
  \item{alpha}{List where rho[[k]][[b]] indicated the
  parameters of the independence distribution in the block
  b of the class k.} \item{proportions}{Vector of classes
  proportions.} }
}

